+++ packages/updating-element/src/updating-element.ts
@@ -219,9 +219,6 @@ const defaultPropertyDeclaration: PropertyDeclaration = {
   hasChanged: notEqual,
 };
 
-export type connectCallback = () => void;
-export type updateCallback = (changedProperties: PropertyValues) => void;
-
 /**
  * The Closure JS Compiler doesn't currently have good support for static
  * property semantics where "this" is dynamic (e.g.
@@ -537,26 +534,6 @@ export abstract class UpdatingElement extends HTMLElement {
    */
   private _reflectingProperty: PropertyKey | null = null;
 
-  /**
-   * Set of callbacks called in connectedCallback.
-   */
-  connectedCallbacks: Set<connectCallback> = new Set();
-
-  /**
-   * Set of callbacks called in disconnectedCallback.
-   */
-  disconnectedCallbacks: Set<connectCallback> = new Set();
-
-  /**
-   * Set of callbacks called before update.
-   */
-  updateCallbacks: Set<updateCallback> = new Set();
-
-  /**
-   * Set of callbacks called after updated.
-   */
-  updatedCallbacks: Set<updateCallback> = new Set();
-
   constructor() {
     super();
     this._updatePromise = new Promise((res) => (this.enableUpdating = res));
@@ -618,8 +595,13 @@ export abstract class UpdatingElement extends HTMLElement {
    */
   connectedCallback() {
     this.enableUpdating();
-    this.connectedCallbacks.forEach((cb) => cb());
+    this._connectedCallback();
   }
+
+  // Note, this is an override point for controller callbacks. Per spec, the
+  // `connectedCallback` itself cannot be changed.
+  //@internal
+  _connectedCallback() {}
 
   /**
    * Note, this method should be considered final and not overridden. It is
@@ -634,8 +616,13 @@ export abstract class UpdatingElement extends HTMLElement {
    * when disconnecting at some point in the future.
    */
   disconnectedCallback() {
-    this.disconnectedCallbacks.forEach((cb) => cb());
+    this._disconnectedCallback();
   }
+
+  // Note, this is an override point for controller callbacks. Per spec, the
+  // `disconnectedCallback` itself cannot be changed.
+  //@internal
+  _disconnectedCallback() {}
 
   /**
    * Synchronizes property values when attributes change.
@@ -766,8 +753,11 @@ export abstract class UpdatingElement extends HTMLElement {
       // This `await` also ensures that property changes are batched.
       await this._updatePromise;
     } catch (e) {
-      // Ignore any previous errors. We only care that the previous cycle is
-      // done. Any error should have been handled in the previous update.
+      // Refire any previous errors async so they do not disrupt the update
+      // cycle. Errors are refired so developers have a chance to observe
+      // them, and this can be done by implementing
+      // `window.onunhandledrejection`.
+      Promise.reject(e);
     }
     const result = this.performUpdate();
     // If `performUpdate` returns a Promise, we await it. This is done to
@@ -821,7 +811,7 @@ export abstract class UpdatingElement extends HTMLElement {
     try {
       shouldUpdate = this.shouldUpdate(changedProperties);
       if (shouldUpdate) {
-        this.updateCallbacks.forEach((cb) => cb(changedProperties));
+        this._willUpdate(changedProperties);
         this.update(changedProperties);
       } else {
         this._markUpdated();
@@ -840,14 +830,18 @@ export abstract class UpdatingElement extends HTMLElement {
     }
   }
 
-  // Note, this is an override point for platform-support.
-  private _afterUpdate(changedProperties: PropertyValues) {
+  // @internal
+  // Note, this is an override point for controller callbacks
+  _willUpdate(_changedProperties: PropertyValues) {}
+
+  // Note, this is an override point for platform-support and controllers.
+  // @internal
+  _afterUpdate(changedProperties: PropertyValues) {
     if (!this.hasUpdated) {
       this.hasUpdated = true;
       this.firstUpdated(changedProperties);
     }
     this.updated(changedProperties);
-    this.updatedCallbacks.forEach((cb) => cb(changedProperties));
   }
 
   private _markUpdated() {