+++ packages/labs/CHANGELOG.md
@@ -22,4 +22,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Adds `UpdatingController`, a class which can interact with an `UpdatingElement` by hooking into its lifecycle, including connected, disconnected, update, updated, and requestUpdate.
+- Adds the callbacks module which provides support for adding lifecycle callbacks for connected, disconnected, update, and updated to UpdatingElement.
 - Adds `labs` package. The `labs` package is an experimental playground for Lit helpers and tools.